planogram-ai-service/requirements.txt
torch==1.1.0
numpy==1.17.4
opencv-python==4.1.2.30
Flask==1.1.1
marshmallow==3.3.0
webargs==5.5.2
pillow==6.2.1
flask_apispec==0.8.4
addict==2.2.1
chardet==3.0.4
cycler==0.10.0
protobuf==3.9.0
requests==2.22.0
tensorboardX==1.8
torchvision==0.3.0
tqdm==4.32.2
aug>=0.1.0
schedule==0.6.0
rtsp==1.1.6
mmcv==0.2.10
face-alignment
keras==2.2.4
